Add new parameter "phy_is_critical" to skip detection of the PHY version.
There are 2 netX90B models which differ in the version of the PHY. They
can only be distinguished by reading out the PHY revision register.
Unfortunately this disturbs a running Ethernet communication. The new
parameter allows tests using Ethernet for communication with the DUT
to skip detection of the PHY version.